requirements.txt
pandas
xlrd
lxml
cssselect
requests
us
namestand
python-dateutil
ipython[all]
